requirements.txt
# 🤖 AI 창업 지원사업 모니터링 시스템 - 의존성 패키지
# 최종 업데이트: 2025.01.03

# ==========================================
# 🌐 웹 프레임워크
# ==========================================
Flask==3.0.0
Flask-CORS==4.0.0

# ==========================================
# 🗄️ 데이터베이스
# ==========================================
supabase>=2.15.0
psycopg2-binary==2.9.9

# ==========================================
# 🔧 환경설정 및 유틸리티
# ==========================================
python-dotenv>=1.0.0
python-dateutil==2.8.2

# ==========================================
# 🌐 웹 크롤링 및 HTTP 통신
# ==========================================
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
httpx>=0.27.2
chardet>=5.0.0  # 인코딩 감지
aiohttp>=3.8.0  # 비동기 HTTP 클라이언트

# ==========================================
# 📊 데이터 처리 및 분석
# ==========================================
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

# ==========================================
# 🤖 AI/ML 라이브러리
# ==========================================
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.2.0
tokenizers>=0.13.0
joblib>=1.2.0

# ==========================================
# 🎯 강화학습 라이브러리
# ==========================================
gymnasium>=1.0.0
stable-baselines3>=1.6.0
optuna>=3.0.0

# ==========================================
# 📈 시각화 및 모니터링
# ==========================================
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0

# ==========================================
# 🔧 기타 유틸리티
# ==========================================
psutil>=5.9.0

# ==========================================
# 🚨 중요 버전 정보
# ==========================================
# Python: 3.8+ 권장 (3.11도 지원)
# CUDA: 11.7+ (GPU 사용시)
# Apple Silicon: MPS 지원

# ==========================================
# 💡 설치 가이드
# ==========================================
# 1. 로컬 전체 설치:
#    pip install -r requirements.txt
#
# 2. 코랩에서 누락된 것만 설치:
#    !pip install scikit-learn stable-baselines3 optuna
#    (코랩 기존 버전과 호환됨)
#
# 3. GPU 버전 (NVIDIA) - 안정성 우선:
#    pip install torch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 --index-url https://download.pytorch.org/whl/cu118
#
# 4. Apple Silicon (M1/M2):
#    기본 설치로 MPS 자동 지원
#
# 5. CPU 전용:
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu

# ==========================================
# 🎯 핵심 기능별 의존성 매핑
# ==========================================
# 웹 대시보드: Flask, Jinja2
# 데이터베이스: supabase, psycopg2-binary
# 웹 크롤링: requests, aiohttp, beautifulsoup4
# AI 분석: torch, transformers, sentence-transformers, scikit-learn
# 강화학습: gym, stable-baselines3, optuna
# 데이터 처리: pandas, numpy
# 시각화: matplotlib, seaborn
# 환경설정: python-dotenv

# ==========================================
# 🔄 코랩/로컬 호환성 노트
# ==========================================
# 코랩에서 이미 설치된 라이브러리들 (더 높은 버전):
# - torch 2.6.0+cu124 (호환됨: >=1.12.0 요구)
# - transformers 4.52.4 (호환됨: >=4.20.0 요구)
# - sentence-transformers 4.1.0 (호환됨: >=2.2.0 요구)
# - numpy 2.0.2 (호환됨: >=1.21.0 요구)
# - pandas 2.2.2 (호환됨: >=1.5.0 요구)
# - matplotlib 3.10.0 (호환됨: >=3.5.0 요구)
# - seaborn 0.13.2 (호환됨: >=0.11.0 요구)
# - tqdm 4.67.1 (호환됨: >=4.64.0 요구)
#
# 로컬에서는 안정적인 버전이 설치되고,
# 코랩에서는 기존 최신 버전이 그대로 사용됩니다.

# ==========================================
# 🎨 테스트된 안정 버전 조합 (참고용)
# ==========================================
# torch==2.0.1
# transformers==4.30.0
# sentence-transformers==2.2.2
# numpy==1.24.3
# pandas==2.0.3
# scikit-learn==1.3.0
# matplotlib==3.7.1
# seaborn==0.12.2 